pyproject.toml
[tool.poetry]
name = "financial-news-sentiment"
version = "0.1.0"
description = "Financial News Sentiment Analysis for Stock Price Prediction"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^1.5.3"
numpy = "^1.24.3"
scikit-learn = "^1.3.0"
yfinance = "^0.2.0"
requests = "^2.28.0"
nltk = "^3.7"
transformers = "^4.21.0"
torch = "^1.12.0"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
textblob = "^0.17.0"
vaderSentiment = "^3.3.2"
sentence-transformers = "^2.2.0"
optuna = "^3.3.0"
xgboost = "^1.7.5"
plotly = "^5.10.0"
flask = "^2.2.0"
fastapi = "^0.85.0"
uvicorn = "^0.18.0"
python-dotenv = "^0.19.0"
pytz = "^2022.1"
ta = "^0.10.0"
jupyter = "^1.0.0"
pyyaml = "^6.0"
dynaconf = "^3.2.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
flake8 = "^6.0.0"
black = "^23.7.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
